refactor usb_camera and apriltags_cuda to be more test friendly
