+++ src/apriltags_cuda/include/apriltags_cuda/BooleanValueSender.h
