+++ src/apriltags_cuda/include/apriltags_cuda/DoubleValueSender.h
