+++ src/apriltags_cuda/include/apriltags_cuda/cameraexception.h
