+++ src/apriltags_cuda/include/apriltags_cuda/labeling_allegretti_2019_BKE.h
