+++ src/apriltags_cuda/src/IntegerArraySender.cpp
@@ -1,6 +1,6 @@
-#include "IntegerArraySender.h"
+#include "apriltags_cuda/IntegerArraySender.h"
 
-#include "NetworkTablesConfig.h"
+#include "apriltags_cuda/NetworkTablesConfig.h"
 
 /*
 Right now, the table is /SmartDashboard so we can visualize the values sent from
