+++ src/apriltags_cuda/src/labeling_allegretti_2019_BKE.cu
@@ -20,7 +20,7 @@
 #include "cuda_runtime.h"
 #include "device_launch_parameters.h"
 #include "glog/logging.h"
-#include "labeling_allegretti_2019_BKE.h"
+#include "apriltags_cuda/labeling_allegretti_2019_BKE.h"
 
 #define BLOCK_ROWS 16
 #define BLOCK_COLS 16
